metrics: add map count metric
Add a new metric to keep track of how many copies of a sensor map are loaded into the
kernel. This value should always be 1 for every sensor map in a properly functioning
Tetragon deployment.

Note that we explicitly target only maps that have been registered to one or more sensors.
We have other maps which are not registered to any sensors but these are generally used
for heap allocation and do not store or share any meaningful data between programs, so it
is okay for them to be duplicated.

Having this as a metric can help us keep track of errors that crop up due to invalid
sensor loading/unloading logic.

Signed-off-by: William Findlay <will@isovalent.com>