tetragon: Create msg handler and remove per type handlers
processManager and eventCache have per type switch statements that
cut'n'paste the logic through to each type. Instead lets create
a interface for message handlers and then we can avoid caring about
event specifics in otherwise generic code.

The hope is this lets us abstract the specifics of any given event
out of the various caches and core logic we maintain. And as a
result we also can drop some open coded interface{} types with
more specific types.

This also gets us closer to allowing stacked sensors.

One point worth noting is that we the New() constructors for the
grpc/* are getting a bit strange now that the grpc object is not
used in Handlers() we might just push those into the __init func
and then we can skip the creator. I don't have any plans or
reason to support multiple constructors on the same sensor.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>