Change eventcache GC run interval
Event cache every "eventRetryTimer" seconds runs a GC (i.e. check all
elements inside the cache if the required info exists). If an element
does not found the required info it is placed again back to the cache.
This happens up to 3 times. After that it generates an event with only
the information that we already have.

Now "eventRetryTimer" is 10 seconds which means:
1. We wait up to 30 seconds to get the required data
2. When we have an OOO exit and exec event, we will get the second event in at max 10 seconds.

This commits reduces the "eventRetryTimer" from 10 to 1 seconds. With
3 retries this will be 3 seconds. If we don't get the info we need after
3 seconds it is fine to stop trying.

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>