Add tests for pkg/grpc/exec/
An exec and exit event can be delivered out of order to the agent.
For this reason, we have a cache that keeps the un-populated events
for a small period of time to wait for the exec event.

There are also other caches that deal with this situations.

This commit adds tests for pkg/grpc/exec. For now it only contains
2 different cases (1) 2 events in order and (2) 2 events out of order.
On the other hand, this can be uses as a skeleton to implement tests
for other caches as well.

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>