Change GetProcessExit to handle out of order events
This commit tries to fix the issues with the previous tests.

In the case of an exit event, if the required data is not fount,
we cache this event for the first time. The second time that we
cannot find the requied data, we return an event with the data
that we already have (i.e. pid and start time).

Signed-off-by: Anastasios Papagiannis <tasos.papagiannnis@gmail.com>